+++ libewf/libewf_endian.c
@@ -39,7 +39,7 @@
 /* Converts a byte array into a 16bit little endian value
  * Returns 1 if successful, -1 on error
  */
-int libewf_endian_convert_16bit( uint16_t *value, uint8_t *bytes )
+int8_t libewf_endian_convert_16bit( uint16_t *value, uint8_t *bytes )
 {
 	static char *function = "libewf_endian_convert_16bit";
 
@@ -67,7 +67,7 @@ int libewf_endian_convert_16bit( uint16_t *value, uint8_t *bytes )
 /* Converts a byte array into a 32bit little endian value
  * Returns 1 if successful, -1 on error
  */
-int libewf_endian_convert_32bit( uint32_t *value, uint8_t *bytes )
+int8_t libewf_endian_convert_32bit( uint32_t *value, uint8_t *bytes )
 {
 	static char *function = "libewf_endian_convert_32bit";
 
@@ -99,7 +99,7 @@ int libewf_endian_convert_32bit( uint32_t *value, uint8_t *bytes )
 /* Converts a byte array into a 64bit little endian value
  * Returns 1 if successful, -1 on error
  */
-int libewf_endian_convert_64bit( uint64_t *value, uint8_t *bytes )
+int8_t libewf_endian_convert_64bit( uint64_t *value, uint8_t *bytes )
 {
 	static char *function = "libewf_endian_convert_64bit";
 
@@ -139,7 +139,7 @@ int libewf_endian_convert_64bit( uint64_t *value, uint8_t *bytes )
 /* Reverts a 16bit value into a little endian byte array
  * Returns 1 if successful, -1 on error
  */
-int libewf_endian_revert_16bit( uint16_t value, uint8_t *bytes )
+int8_t libewf_endian_revert_16bit( uint16_t value, uint8_t *bytes )
 {
 	static char *function = "libewf_endian_revert_16bit";
 
@@ -160,7 +160,7 @@ int libewf_endian_revert_16bit( uint16_t value, uint8_t *bytes )
 /* Reverts a 32bit value into a little endian byte array
  * Returns 1 if successful, -1 on error
  */
-int libewf_endian_revert_32bit( uint32_t value, uint8_t *bytes )
+int8_t libewf_endian_revert_32bit( uint32_t value, uint8_t *bytes )
 {
 	static char *function = "libewf_endian_revert_32bit";
 
@@ -185,7 +185,7 @@ int libewf_endian_revert_32bit( uint32_t value, uint8_t *bytes )
 /* Reverts a 64bit value into a little endian byte array
  * Returns 1 if successful, -1 on error
  */
-int libewf_endian_revert_64bit( uint64_t value, uint8_t *bytes )
+int8_t libewf_endian_revert_64bit( uint64_t value, uint8_t *bytes )
 {
 	static char *function = "libewf_endian_revert_64bit";
